Fix overloaded funtion to call I2C master begin

+++ src/M5_RoverC.cpp
@@ -6,7 +6,7 @@ bool M5_RoverC::begin(TwoWire *wire, uint8_t sda, uint8_t scl, uint8_t addr) {
     _addr = addr;
     _sda  = sda;
     _scl  = scl;
-    _wire->begin(_sda, _scl);
+    _wire->begin((int)_sda, (int)_scl);
     delay(10);
     _wire->beginTransmission(_addr);
     uint8_t error = _wire->endTransmission();
